@@ -1,10 +1,9 @@
-import Vue from 'vue';
+import { createApp } from 'vue';
 import Element from 'main/index.js';
 import App from './play/index.vue';
 import 'packages/theme-chalk/src/index.scss';
 
-Vue.use(Element);
+const app = createApp(App);
+app.use(Element);
+app.mount('#app');
 
-new Vue({ // eslint-disable-line
-  render: h => h(App)
-}).$mount('#app');
