@@ -1,8 +1,8 @@
 import ElButton from './src/button';
 
 /* istanbul ignore next */
-ElButton.install = function(Vue) {
-  Vue.component(ElButton.name, ElButton);
+ElButton.install = function(app) {
+  app.component(ElButton.name, ElButton);
 };
 
 export default ElButton;
