@@ -1,8 +1,6 @@
 /* istanbul ignore next */
 
-import Vue from 'vue';
-
-const isServer = Vue.prototype.$isServer;
+const isServer = false;
 const SPECIAL_CHARS_REGEXP = /([\:\-\_]+(.))/g;
 const MOZ_HACK_REGEXP = /^moz([A-Z])/;
 const ieVersion = isServer ? 0 : Number(document.documentMode);
